requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
greenlet==3.0.3
grpcio==1.65.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
protobuf==4.25.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
tenacity==9.0.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b105
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
